Extracting `is_expr_identity_function` into `clippy_utils` for reusability
